[Form] Compound forms now always need a data mapper. Otherwise an exception is thrown.
