+++ src/Symfony/Component/Form/Tests/AbstractFormTest.php
@@ -1,0 +1,141 @@
+<?php
+
+    /*
+    * This file is part of the Symfony package.
+    *
+    * (c) Fabien Potencier <fabien@symfony.com>
+    *
+    * For the full copyright and license information, please view the LICENSE
+    * file that was distributed with this source code.
+    */
+
+namespace Symfony\Component\Form\Tests;
+
+use Symfony\Component\Form\FormError;
+use Symfony\Component\Form\FormBuilder;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
+
+abstract class AbstractFormTest extends \PHPUnit_Framework_TestCase
+{
+    /**
+     * @var EventDispatcherInterface
+     */
+    protected $dispatcher;
+
+    /**
+     * @var \Symfony\Component\Form\FormFactoryInterface
+     */
+    protected $factory;
+
+    /**
+     * @var \Symfony\Component\Form\FormInterface
+     */
+    protected $form;
+
+    protected function setUp()
+    {
+        if (!class_exists('Symfony\Component\EventDispatcher\EventDispatcher')) {
+            $this->markTestSkipped('The "EventDispatcher" component is not available');
+        }
+
+        $this->dispatcher = $this->getMock('Symfony\Component\EventDispatcher\EventDispatcherInterface');
+        $this->factory = $this->getMock('Symfony\Component\Form\FormFactoryInterface');
+        $this->form = $this->createForm();
+    }
+
+    protected function tearDown()
+    {
+        $this->dispatcher = null;
+        $this->factory = null;
+        $this->form = null;
+    }
+
+    /**
+     * @return \Symfony\Component\Form\FormInterface
+     */
+    abstract protected function createForm();
+
+    /**
+     * @param string                   $name
+     * @param EventDispatcherInterface $dispatcher
+     * @param string                   $dataClass
+     *
+     * @return FormBuilder
+     */
+    protected function getBuilder($name = 'name', EventDispatcherInterface $dispatcher = null, $dataClass = null)
+    {
+        return new FormBuilder($name, $dataClass, $dispatcher ?: $this->dispatcher, $this->factory);
+    }
+
+    /**
+     * @param  string $name
+     *
+     * @return \PHPUnit_Framework_MockObject_MockObject
+     */
+    protected function getMockForm($name = 'name')
+    {
+        $form = $this->getMock('Symfony\Component\Form\Tests\FormInterface');
+
+        $form->expects($this->any())
+            ->method('getName')
+            ->will($this->returnValue($name));
+
+        return $form;
+    }
+
+    /**
+     * @param  string $name
+     *
+     * @return \PHPUnit_Framework_MockObject_MockObject
+     */
+    protected function getValidForm($name)
+    {
+        $form = $this->getMockForm($name);
+
+        $form->expects($this->any())
+            ->method('isValid')
+            ->will($this->returnValue(true));
+
+        return $form;
+    }
+
+    /**
+     * @param  string $name
+     *
+     * @return \PHPUnit_Framework_MockObject_MockObject
+     */
+    protected function getInvalidForm($name)
+    {
+        $form = $this->getMockForm($name);
+
+        $form->expects($this->any())
+            ->method('isValid')
+            ->will($this->returnValue(false));
+
+        return $form;
+    }
+
+    /**
+     * @return \PHPUnit_Framework_MockObject_MockObject
+     */
+    protected function getDataMapper()
+    {
+        return $this->getMock('Symfony\Component\Form\DataMapperInterface');
+    }
+
+    /**
+     * @return \PHPUnit_Framework_MockObject_MockObject
+     */
+    protected function getDataTransformer()
+    {
+        return $this->getMock('Symfony\Component\Form\DataTransformerInterface');
+    }
+
+    /**
+     * @return \PHPUnit_Framework_MockObject_MockObject
+     */
+    protected function getFormValidator()
+    {
+        return $this->getMock('Symfony\Component\Form\FormValidatorInterface');
+    }
+}
