+++ src/Symfony/Component/Form/Tests/Extension/Core/DataMapper/PropertyPathMapperTest.php
@@ -71,6 +71,14 @@ class PropertyPathMapperTest extends \PHPUnit_Framework_TestCase
         return $form;
     }
 
+    /**
+     * @return \PHPUnit_Framework_MockObject_MockObject
+     */
+    private function getDataMapper()
+    {
+        return $this->getMock('Symfony\Component\Form\DataMapperInterface');
+    }
+
     public function testMapDataToFormsPassesObjectRefIfByReference()
     {
         $car = new \stdClass();
@@ -181,6 +189,8 @@ class PropertyPathMapperTest extends \PHPUnit_Framework_TestCase
         $config = new FormConfig('name', '\stdClass', $this->dispatcher);
         $config->setByReference(true);
         $config->setVirtual(true);
+        $config->setCompound(true);
+        $config->setDataMapper($this->getDataMapper());
         $form = $this->getForm($config);
 
         $config = new FormConfig('engine', '\stdClass', $this->dispatcher);
@@ -350,6 +360,8 @@ class PropertyPathMapperTest extends \PHPUnit_Framework_TestCase
         $config = new FormConfig('name', '\stdClass', $this->dispatcher);
         $config->setPropertyPath($parentPath);
         $config->setVirtual(true);
+        $config->setCompound(true);
+        $config->setDataMapper($this->getDataMapper());
         $form = $this->getForm($config);
 
         $config = new FormConfig('engine', '\stdClass', $this->dispatcher);
